test(e2e): isolate VM-specific smoke assertions
